refactor: rename bulkCodeConstants.ts
Revert "refactor: rename bulkCodeConstants.ts"

This reverts commit c49be9014abbb307227799552868fe4740fa4fef.

refactor: rename bulkCodeConstants.ts
