+++ pages/profile/bulk-codes/[method]/[id].tsx
@@ -4,7 +4,7 @@ import BulkCodes from '../../../../src/features/user/BulkCodes';
 import { PaymentOptions } from '../../../../src/features/user/BulkCodes/BulkCodesTypes';
 import Head from 'next/head';
 import i18next from '../../../../i18n';
-import { BulkCodeMethods } from '../../../../src/utils/constants/bulkCodeMethods';
+import { BulkCodeMethods } from '../../../../src/utils/constants/bulkCodeConstants';
 import { useBulkCode } from '../../../../src/features/common/Layout/BulkCodeContext';
 import { ErrorHandlingContext } from '../../../../src/features/common/Layout/ErrorHandlingContext';
 import { getRequest } from '../../../../src/utils/apiRequests/api';
