+++ pages/profile/bulk-codes/[method]/index.tsx
@@ -5,7 +5,7 @@ import Head from 'next/head';
 import i18next from '../../../../i18n';
 import { useRouter } from 'next/router';
 import { useBulkCode } from '../../../../src/features/common/Layout/BulkCodeContext';
-import { BulkCodeMethods } from '../../../../src/utils/constants/bulkCodeMethods';
+import { BulkCodeMethods } from '../../../../src/utils/constants/bulkCodeConstants';
 
 const { useTranslation } = i18next;
