+++ src/utils/constants/bulkCodeConstants.ts
